refactor(mobileapp): create job folder to requests folder
