+++ mobileapp/src/common/api/hooks/useJob.ts
@@ -1,7 +1,7 @@
 import { useQuery } from "react-query";
 import type { Job } from "../../../models/job";
 
-import getJob from "../requests/getJob";
+import getJob from "../requests/job/getJob";
 
 export default function useJob(jobId: string) {
   const { data, isLoading, error } = useQuery<Job>(
