+++ mobileapp/src/common/api/hooks/useJobs.ts
@@ -1,7 +1,7 @@
 import { useQuery } from "react-query";
 import type { Job } from "../../../models/job";
 import PaginatedDataType from "../interfaces/PaginatedDataType";
-import getJobs from "../requests/getJobs";
+import getJobs from "../requests/job/getJobs";
 
 export default function useJobs() {
   const { data, isLoading, error } = useQuery<PaginatedDataType<Job>>(
